requirements.txt
numpy>=1.21.0
pandas>=1.3.0
gym>=0.21.0
torch>=1.9.0
websocket-client>=1.2.1
python-dotenv>=0.19.0
requests>=2.26.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
pyarrow>=5.0.0  # for parquet support

